app/worker/requirements.txt
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-json-logger==2.0.7
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
prometheus_client==0.20.0
